rename settings form component
